View the generated password.
As of right now the password get's sent to the appropriate email and that might be enough but we really want to ease the usage of this feature and make it a bit more error ressilient. The new deployment does not seems to let mail work so this is basically an escape hatch.