Fix division by zero error when retrieving entries
If a user does not have any normalized working days in the range we
get a division by zero error.

Not having any normalized working days in a period of time is actually
to be expected for users which are on leave, prolonged vacation or the
like.

To address this we introduce a conditional to check for this case.

We also use this opportunity to introduce intermediary variables with
meaningful names to clarify the calculation.