Validate Opcache based on timestamps in Docker
While not validating timestamps for files may be a nice performance
improvement it makes development a lot more complicated as the cache
will have to be reset for each file change.

In practice this means restarting Apache in the container or restarting
the entire container.

This is more troublesome than the value provided by the performance
improvement and we have no easy way to toggle between this between
development and production.

Consequently we reenable Opcache timestamp validation across 
environments.